% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/askme.R
\name{askme_cancertype}
\alias{askme_cancertype}
\title{classify cancer type}
\usage{
askme_cancertype(betas, model = m_cancertype_TCGA33)
}
\arguments{
\item{betas}{DNA methylation betas}

\item{model}{cancer classification model}
}
\value{
predicted cancer type label
}
\description{
classify cancer type
}
\examples{
library(sesameData)
betas <- sesameDataGet("HM450.1.TCGA.PAAD")$betas
askme_cancertype(betas, model=m_cancertype_TCGA33)
## expect PAAD
}
