% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/askme.R
\name{askme_classify}
\alias{askme_classify}
\title{classify sample.}
\usage{
askme_classify(betas, model)
}
\arguments{
\item{model}{classification model}

\item{sample}{DNA methylation beta}
}
\value{
predicted cancer type label
}
\description{
classify sample.
}
\examples{
library(sesameData)
betas <- sesameDataGet("HM450.1.TCGA.PAAD")$betas
askme_classify(betas, m_cancertype_TCGA33)
## expect PAAD
}
