% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/askme.R
\docType{data}
\name{askme_models}
\alias{askme_models}
\title{Master data frame for all model objects}
\value{
master sheet of askme model objects
}
\description{
This is an internal object which will be updated on every new release
}
\examples{
print(askme_models$ModelID)
}
