% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytomethic.R
\name{cmi_checkVersion}
\alias{cmi_checkVersion}
\title{Check CytoMethIC versions}
\usage{
cmi_checkVersion()
}
\value{
print the versions of cytomethic and dependencies
}
\description{
print package verison of cytomethic and depended packages to help
troubleshoot installation issues.
}
\examples{
cmi_checkVersion()
}
