% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytomethic.R
\name{cmi_classify}
\alias{cmi_classify}
\title{The cmi_classify function takes in a model and a sample, and uses the model to classify it.
This function supports randomForest, e1071::svm, xgboost, and keras/tensorflow models. For xgboost and keras models,
the features used in classification as well as a label mapping must be provided for output.}
\usage{
cmi_classify(betas, model, feature = NULL, label_levels = NULL)
}
\arguments{
\item{betas}{DNA methylation beta}

\item{model}{classification model}

\item{feature}{list of features if not stored within model}

\item{label_levels}{factor-label mapping if not stored within model}
}
\value{
predicted cancer type label
}
\description{
The cmi_classify function takes in a model and a sample, and uses the model to classify it.
This function supports randomForest, e1071::svm, xgboost, and keras/tensorflow models. For xgboost and keras models,
the features used in classification as well as a label mapping must be provided for output.
}
\examples{
library(sesameData)
betas <- sesameDataGet("HM450.1.TCGA.PAAD")$betas
cmi_classify(betas, m_cancertype_CNS66)
}
