% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytomethic.R
\name{cmi_classify_rda}
\alias{cmi_classify_rda}
\title{The cmi_classify_rda function takes in a model and a sample, and uses the model to classify it.
This function supports randomForest, e1071::svm, xgboost, and keras/tensorflow models. WARNING: This will only
work with CytoMethIC models on ExperimentHub, or .rda files formatted exactly how the CytoMethIC models are.}
\usage{
cmi_classify_rda(betas, model_path)
}
\arguments{
\item{betas}{DNA methylation beta}

\item{model_path}{Path to the classification model}
}
\value{
predicted cancer type label
}
\description{
The cmi_classify_rda function takes in a model and a sample, and uses the model to classify it.
This function supports randomForest, e1071::svm, xgboost, and keras/tensorflow models. WARNING: This will only
work with CytoMethIC models on ExperimentHub, or .rda files formatted exactly how the CytoMethIC models are.
}
\examples{
library(sesameData)
betas <- sesameDataGet("HM450.1.TCGA.PAAD")$betas
}
