% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytomethic.R
\docType{data}
\name{cmi_models}
\alias{cmi_models}
\title{Master data frame for all model objects}
\value{
master sheet of CytoMethIC model objects
}
\description{
This is an internal object which will be updated on every new release
}
\examples{
print(cmi_models[,c("EHID","Title")])
}
