% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{.optimizeFrac}
\alias{.optimizeFrac}
\title{Internal function for fraction optimization}
\usage{
.optimizeFrac(
  frac,
  ref,
  q,
  errFunc,
  temp = 0.5,
  maxIter = 1000,
  delta = 1e-04,
  step.max = 1,
  verbose = FALSE
)
}
\arguments{
\item{frac}{initial fraction}

\item{ref}{reference}

\item{q}{query}

\item{errFunc}{error function}

\item{temp}{annealing temperature}

\item{maxIter}{maximum iteration to stop after converge}

\item{delta}{delta score to reset counter}

\item{step.max}{maximum step, do not adjust}

\item{verbose}{output debug info}
}
\value{
a list of fractions and min err
}
\description{
Internal function for fraction optimization
}
