% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytomethic.R
\name{impute_mean}
\alias{impute_mean}
\title{Impute Missing Values with Mean
This function replaces missing values (NA) in a data frame or matrix, default is col means.}
\usage{
impute_mean(df, axis = 1)
}
\arguments{
\item{df}{A dataframe or matrix}

\item{axis}{A single integer. Use 1 to impute column means (default), and 2 to impute row means.}
}
\value{
A data frame or matrix with missing values imputed.
}
\description{
Impute Missing Values with Mean
This function replaces missing values (NA) in a data frame or matrix, default is col means.
}
\examples{
df <- data.frame(a = c(1, 2, NA, 4), b = c(NA, 2, 3, 4))
df <- impute_mean(df)
df <- impute_mean(df, axis = 2)
}
