% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytomethic.R
\name{impute_mean_cmi}
\alias{impute_mean_cmi}
\title{Impute Missing Values with Mean, for CytoMethIC models ONLY
This function replaces missing values (NA) in a data frame or matrix, default is col means.}
\usage{
impute_mean_cmi(df)
}
\arguments{
\item{df}{A dataframe or matrix}
}
\value{
A data frame or matrix with missing values imputed.
}
\description{
Impute Missing Values with Mean, for CytoMethIC models ONLY
This function replaces missing values (NA) in a data frame or matrix, default is col means.
}
