% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytomethic.R
\name{inferSex}
\alias{inferSex}
\title{Infer sex.}
\usage{
inferSex(betas)
}
\arguments{
\item{betas}{DNA methylation beta}
}
\value{
Inferred sex of sample
}
\description{
Infer sex.
}
\examples{
library(sesameData)
betas <- sesameDataGet("HM450.1.TCGA.PAAD")$betas
inferSex(betas)
}
