% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytomethic.R
\docType{data}
\name{m_cancertype_CNS66}
\alias{m_cancertype_CNS66}
\title{Model for CNS cancer type classification}
\value{
random forest model
}
\description{
Model for CNS cancer type classification
}
\examples{
print(m_cancertype_CNS66$ntree)
}
