% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/askme.R
\docType{data}
\name{m_cancertype_TCGA33}
\alias{m_cancertype_TCGA33}
\title{Model for TCGA cancer type classification}
\value{
random forest model
}
\description{
Model for TCGA cancer type classification
}
\examples{
print(m_cancertype_TCGA33$ntree)
}
