% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytomethic.R
\docType{data}
\name{prediction_labels}
\alias{prediction_labels}
\title{Master data frame for all prediction labels}
\value{
master sheet of prediction labels
}
\description{
This is an internal object which will be updated on every new release
}
\examples{
print(prediction_labels)
}
