% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/askme.R
\name{askme_classify}
\alias{askme_classify}
\title{classify sample.}
\usage{
askme_classify(sample, model, Probe_IDs)
}
\arguments{
\item{sample}{DNA methylation beta}

\item{model}{classification model}

\item{Probe_IDs}{methylation probes used}
}
\value{
predicted cancer type label
}
\description{
classify sample.
}
\examples{
library(sesameData)
betas <- sesameDataGet("HM450.1.TCGA.PAAD")$betas
model <- readRDS("rfc_model.rds")
probes <- readRDS("HM450.rds")
askme_classify(betas[1], model, probes)
## expect PAAD
}
