% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/askme.R
\name{askme_classify}
\alias{askme_classify}
\title{classify sample.}
\usage{
askme_classify(betas, model, feature = NULL, label_levels = NULL)
}
\arguments{
\item{betas}{DNA methylation beta}

\item{model}{classification model}

\item{feature}{list of features if not stored within model}

\item{label_levels}{factor-label mapping if not stored within model}
}
\value{
predicted cancer type label
}
\description{
classify sample.
}
\examples{
library(sesameData)
betas <- sesameDataGet("HM450.1.TCGA.PAAD")$betas
askme_classify(betas, m_cancertype_CNS66)
## expect PAAD
}
