% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytomethic.R
\name{cmi_classify}
\alias{cmi_classify}
\title{The cmi_classify function takes in a model and a sample, and uses the model to classify it.
This function supports randomForest, e1071::svm, xgboost, and keras/tensorflow models. For xgboost and keras models,
the features used in classification as well as a label mapping must be provided for output.}
\usage{
cmi_classify(betas, cmi_model)
}
\arguments{
\item{betas}{DNA methylation beta}

\item{cmi_model}{Cytomethic model downloaded from ExperimentHub}
}
\value{
predicted cancer type label
}
\description{
The cmi_classify function takes in a model and a sample, and uses the model to classify it.
This function supports randomForest, e1071::svm, xgboost, and keras/tensorflow models. For xgboost and keras models,
the features used in classification as well as a label mapping must be provided for output.
}
\examples{
library(sesameData)
library(ExperimentHub)
eh <- ExperimentHub()
model <- eh[["EH8395"]]
betas <- sesameDataGet("HM450.1.TCGA.PAAD")$betas
cmi_classify(betas, model)
#Expect PAAD
}
