% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytomethic.R
\name{cmi_predict}
\alias{cmi_predict}
\title{The cmi_predict function takes in a model and a sample, and uses the model
to classify it.  This function supports randomForest, e1071::svm, xgboost,
and keras/tensorflow models. For xgboost and keras models, the features used
in classification as well as a label mapping must be provided for output.}
\usage{
cmi_predict(
  betas,
  cmi_model,
  source_platform = NULL,
  lift_over = FALSE,
  verbose = FALSE,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{betas}{DNA methylation beta}

\item{cmi_model}{Cytomethic model downloaded from ExperimentHub}

\item{source_platform}{source platform
If not given, will infer from probe ID.}

\item{lift_over}{whether to allow mLiftOver to convert probe IDs}

\item{verbose}{be verbose with warning}

\item{BPPARAM}{use MulticoreParam(n) for parallel processing}
}
\value{
predicted cancer type label
}
\description{
The cmi_predict function takes in a model and a sample, and uses the model
to classify it.  This function supports randomForest, e1071::svm, xgboost,
and keras/tensorflow models. For xgboost and keras models, the features used
in classification as well as a label mapping must be provided for output.
}
\examples{

library(sesame)
library(ExperimentHub)
library(CytoMethIC)

## Cancer Type
model = ExperimentHub()[["EH8395"]]
cmi_predict(openSesame(sesameDataGet("EPICv2.8.SigDF")[[1]]), model, lift_over=TRUE)
cmi_predict(openSesame(sesameDataGet('EPIC.1.SigDF')), model, lift_over=TRUE)
cmi_predict(sesameDataGet("HM450.1.TCGA.PAAD")$betas, model, lift_over=TRUE)

}
